% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate_binary_outcome}
\alias{associate_binary_outcome}
\title{Associate GRPs with a binary outcome}
\usage{
associate_binary_outcome(
  decomp,
  metadata,
  outcome.name,
  other.predictors = NULL,
  genelists = NULL,
  run.ML = T,
  cores = 6,
  replicates = 3,
  use.s = "lambda.min"
)
}
\arguments{
\item{decomp}{Object of class \code{\link{decomposition}}, output from \code{\link{decompose}}}

\item{metadata}{A data frame with rownames corresponding to column names of \code{decomp} that contains the outcome of interest and any other predictors that should be tested}

\item{outcome.name}{Column name from \code{metadata} that contains the outcome of interest}

\item{other.predictors}{Names of any other columns in \code{metadata} that may serve as 0-shot predictors for the outcome}

\item{genelists}{List of character vectors of gene names. If \code{run.ML=TRUE}, LASSO regression using these genes as input will be performed and compared to the other models.}

\item{run.ML}{Whether to run machine-learning models.}

\item{cores}{Number of cores to use (if replicates > 1)}

\item{replicates}{Number of times to repeat CV calculation with different train-test splits. Only used if \code{run.ML=TRUE}}

\item{use.s}{Regularization to use for LASSO, one of \code{lambda.min} and \code{lambda.1se}}
}
\value{
A list with entries \code{p.values} and, if \code{run.ML} is set, \code{aucs}, their standard deviation across replicates, and info on the different folds
}
\description{
This function takes a decomposition object and a metadata frame with a binary (TRUE/FALSE) outcome and, if available, other predictors. It will then a) compute statistical associations between
all predictors and the outcome and b) compare the performance of all predictors to that of various supervised machine learning strategies (see details)
}
\details{
If \code{run.ML} is set, the function will set up a 10 fold cross validation scheme (or leave one out CV if <50 observations). On the train set, it will perform the following:
\itemize{
\item{Perform principal component regression on 18 PCs computed on highly variable genes from the underlying Seurat object}
\item{Perform LASSO regularized principal component regression on 50 PCs computed on highly variable genes from the underlying Seurat object}
\item{Perform DE testing between samples from the highest and lowest quartile of the outcome, and perform LASSO or linear regression on the top DE features}
\item{Perform LASSO and linear regression on the 18 GRPs from the decomposition}
\item{If \code{genelists} are provided, perform LASSO on the genes from the gene lists}
}
}
