% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition-class.R
\name{colnames,decomposition-method}
\alias{colnames,decomposition-method}
\title{Get column names of a decomposition object}
\usage{
\S4method{colnames}{decomposition}(x, do.NULL = TRUE, prefix = "col")
}
\arguments{
\item{x}{A \code{\link{decomposition}} object.}

\item{do.NULL, }{prefix Passed to \code{\link[base]{colnames}}.}
}
\value{
A character vector of column names.
}
\description{
Get column names of a decomposition object
}
