% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition-class.R
\name{colnames<-,decomposition,character-method}
\alias{colnames<-,decomposition,character-method}
\title{Set column names of a decomposition object}
\usage{
\S4method{colnames}{decomposition,character}(x) <- value
}
\arguments{
\item{x}{A \code{\link{decomposition}} object.}

\item{value}{A character vector of new column names.}
}
\value{
The updated \code{\link{decomposition}} object.
}
\description{
Set column names of a decomposition object
}
