% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{convertIDs}
\alias{convertIDs}
\alias{convertIDs,matrix,character-method}
\alias{convertIDs,character,character-method}
\title{Convert gene IDs}
\usage{
convertIDs(x, rid)

\S4method{convertIDs}{matrix,character}(x, rid)

\S4method{convertIDs}{character,character}(x, rid)
}
\arguments{
\item{x}{A matrix with gene symbols as row names, or a character vector gene symbols.}

\item{rid}{One of hsy (human gene symbol), hens or mens (ensembl IDs)}
}
\value{
The matrix with ambiguous entries removed, or the character vector
}
\description{
Convert mouse gene symbols to different formats
}
\section{Functions}{
\itemize{
\item \code{convertIDs(x = matrix, rid = character)}: Method for matrices

\item \code{convertIDs(x = character, rid = character)}: Method for character vectors

}}
