% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillaries.R
\name{get_genes}
\alias{get_genes}
\title{Show genes associated with a certain GRP}
\usage{
get_genes(factor, top = T, values = F)
}
\arguments{
\item{factor}{Name of a GRP, e.g. \code{"Factor_7"}}

\item{top}{Whether to show genes with the highest or lowest loadings}

\item{values}{Whether to return numerical factor loadings}
}
\value{
Character vector of gene symbols, sorted by absolute value of the loading, or named numeric vector
}
\description{
Show genes associated with a certain GRP
}
