% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillaries.R
\name{get_pert}
\alias{get_pert}
\title{Show perturbations regulating a certain GRP}
\usage{
get_pert(factor, top = T, values = F)
}
\arguments{
\item{factor}{Name of a GRP, e.g. \code{"Factor_7"}}

\item{top}{Whether to show positive or negative regulators}

\item{values}{Whether to return numerical factor loadings}
}
\value{
Character vector of gene symbols, sorted by absolute value of the loading, or named numeric vector
}
\description{
Show perturbations regulating a certain GRP
}
