% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsc19-package.R
\docType{package}
\name{hsc19-package}
\alias{hsc19}
\alias{hsc19-package}
\title{hsc19: 19 gene regulatory programs in HSCs}
\description{
Decomposes transcriptome data into 19 gene regulatory programs defined by Perturb-Seq
}
\author{
\strong{Maintainer}: Lars Velten \email{lars.velten@crg.eu}

Authors:
\itemize{
  \item Joseph Bowness \email{joseph.bowness@crg.eu}
}

}
\keyword{internal}
