% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillaries.R
\name{plot_pert}
\alias{plot_pert}
\title{Plot perturbations regulating a certain GRP}
\usage{
plot_pert(factor, min.sig.gRNA = 2)
}
\arguments{
\item{factor}{Name of a GRP, e.g. \code{"Factor_7"}}

\item{min.sig.gRNA}{Only include perturbations with this many significant gRNAs}
}
\value{
ggplot2 object
}
\description{
Plot perturbations regulating a certain GRP
}
