% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition-class.R
\name{show,decomposition-method}
\alias{show,decomposition-method}
\title{Show a decomposition object}
\usage{
\S4method{show}{decomposition}(object)
}
\arguments{
\item{object}{A \code{\link{decomposition}} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Print a short summary of a \code{\link{decomposition}}.
}
