% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition-class.R
\name{subset,decomposition-method}
\alias{subset,decomposition-method}
\title{Subset a decomposition object}
\usage{
\S4method{subset}{decomposition}(x, include)
}
\arguments{
\item{x}{A \code{\link{decomposition}} object.}

\item{include}{Samples/cells to keep. This is used to subset \code{x@result[, include]}
and is passed to \code{cells = include} in \code{\link[Seurat:subset.Seurat]{Seurat::subset()}}.}
}
\value{
A \code{\link{decomposition}} object containing only \code{include}.
}
\description{
Keep only a subset of samples/cells in a \code{\link{decomposition}}.
}
\seealso{
\code{\link[base]{subset}}
}
