% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose RNA-seq data into 18 gene regulatory programs}
\usage{
decompose(counts = NULL, s = NULL, loadings = hsc19_loadings)
}
\arguments{
\item{counts}{Count matrix, with gene IDs as row names and sample identifiers as column names. Supported gene name formats are mouse or human gene symbols (preferred), or ENSEMBL (ENSG/ENSMUSG) IDs.}

\item{s}{A seurat object, with \code{NormalizeData} and \code{FindVariableFeatures} run. One of \code{counts} or \code{s} needs to be provided. If \code{counts} is provided, \code{s} is ignored.}

\item{loadings}{Loading matrix with genes in rows and factors/programs in columns. Defaults to \code{hsc19_loadings}.}
}
\value{
Returns an object of class \code{\link{decomposition}}
}
\description{
Ideally run this function on RNA-seq data of HSCs, e.g. bulk RNA-seq data of CD34+ cells (human) or LSK cells (mouse), or single cell data subset to HSCs. This function can run on bulk, pseudobulk or single cell RNA-seq.
}
