% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition-class.R
\docType{class}
\name{decomposition-class}
\alias{decomposition-class}
\alias{decomposition}
\title{Decomposition class}
\description{
This class stores the results of \code{\link{decompose}}
}
\section{Slots}{

\describe{
\item{\code{result}}{inferred activity of the 18 GRPs per sample}

\item{\code{pca}}{result of running `prcomp` on the HVGs of the input data}

\item{\code{pve}}{Percent variance explained}

\item{\code{s}}{Seurat object that was used to create the decomposition}

\item{\code{pve_pca}}{Percent variant explained by 18 principal components}
}}

